p4i integration Integrate 105109 from 3 to main: Integrate 105107 from 3.2 to 3: no compile warnings w/ gcc 3.3
[git-p4: depot-paths = "//depot/qt/main/": change = 105110]

+++ src/widgets/qtoolbox.cpp
@@ -534,7 +534,7 @@ void QToolBox::activateClosestPage( QWidget *widget )
     int curIndexUp = d->pageList.findIndex( *c );
     int curIndexDown = curIndexUp;
     while ( !p ) {
-	if ( curIndexDown < d->pageList.count()-1 ) {
+	if ( curIndexDown < (int)d->pageList.count()-1 ) {
 	    c = d->page(page(++curIndexDown));
 	    if ( c->button->isEnabled() ) {
 		p = c->widget;
